Refactor to move code from utils to the loader, cluster launcher does all restart
