+++ src/timescale_bgw.h
@@ -1,3 +1,3 @@
 
-
-extern void timescale_bgw_scheduler_main(Oid db_id);
+#include <postgres.h>
+extern void timescale_bgw_db_scheduler_main(Oid db_id);
